requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=1.1.0=pypi_0
accelerate=0.13.0=pypi_0
accelerator=2022.8.4.dev1=pypi_0
adapter-transformers=3.1.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
apex=0.1=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
bleach=5.0.1=pypi_0
bottle=0.12.23=pypi_0
bs4=0.0.1=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.9.24=ha878542_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=pypi_0
cffi=1.15.1=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.1.0=pypi_0
click=8.1.3=pypi_0
cycler=0.11.0=pypi_0
d2l=0.17.5=pypi_0
datasets=2.5.1=pypi_0
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.5.1=pypi_0
einops=0.5.0=pypi_0
entrypoints=0.4=pypi_0
et-xmlfile=1.1.0=pypi_0
evaluate=0.2.2=pypi_0
executing=0.8.3=pypi_0
faiss-gpu=1.7.2=pypi_0
fastjsonschema=2.15.3=pypi_0
filelock=3.8.0=pypi_0
fonttools=4.33.3=pypi_0
frozenlist=1.3.0=pypi_0
fsspec=2022.5.0=pypi_0
futures=3.0.5=pypi_0
google-auth=2.9.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
goslate=1.5.4=pypi_0
grpcio=1.47.0=pypi_0
higher=0.2.1=pypi_0
huggingface-hub=0.10.0=pypi_0
hydra-core=1.2.0=pypi_0
idna=2.10=pypi_0
importlib-metadata=4.12.0=pypi_0
importlib-resources=5.8.0=pypi_0
ipykernel=6.15.0=pypi_0
ipython=8.4.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.1=pypi_0
jedi=0.18.1=pypi_0
jieba=0.42.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jsonschema=4.6.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.3.4=pypi_0
jupyter-console=6.4.4=pypi_0
jupyter-core=4.10.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=1.1.1=pypi_0
kiwisolver=1.4.3=pypi_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
libblas=3.9.0=16_linux64_openblas
libcblas=3.9.0=16_linux64_openblas
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=12.1.0=h69a702a_16
libgfortran5=12.1.0=hdcd56e2_16
libgomp=12.1.0=h8d9b700_16
liblapack=3.9.0=16_linux64_openblas
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.21=pthreads_h78a6416_3
libprotobuf=3.20.1=h6239696_4
libstdcxx-ng=12.1.0=ha89aaad_16
libuuid=2.32.1=h7f98852_1000
libzlib=1.2.12=h166bdaf_1
lightly=1.2.22=pypi_0
lightly-utils=0.0.2=pypi_0
loralib=0.1.1=pypi_0
markdown=3.3.7=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
more-itertools=8.14.0=pypi_0
multidict=6.0.2=pypi_0
multiprocess=0.70.13=pypi_0
nbclient=0.6.6=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.4.0=pypi_0
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.5=pypi_0
nltk=3.7=pypi_0
notebook=6.4.12=pypi_0
numpy=1.21.5=pypi_0
oauthlib=3.2.0=pypi_0
omegaconf=2.2.2=pypi_0
openpyxl=3.0.10=pypi_0
openssl=3.0.5=h166bdaf_2
packaging=21.3=pypi_0
pandas=1.2.4=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=pypi_0
pip=22.1.2=pyhd8ed1ab_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.30=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=9.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pydeprecate=0.3.2=pypi_0
pydictionary=2.0.1=pypi_0
pyexecjs=1.5.1=pypi_0
pygments=2.12.0=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.8.13=ha86cf86_0_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch-lightning=1.6.4=pypi_0
pytorch-model-summary=0.1.2=pypi_0
pytz=2022.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=pypi_0
qtconsole=5.3.1=pypi_0
qtpy=2.1.0=pypi_0
rank-bm25=0.2.2=pypi_0
readline=8.1.2=h0f457ee_0
regex=2022.9.13=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.1=pypi_0
send2trash=1.8.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=63.1.0=py38h578d9bd_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0=pypi_0
solidity-parser=0.1.1=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.39.0=h4ff8645_0
stack-data=0.3.0=pypi_0
synonyms=3.18.0=pypi_0
terminado=0.15.0=pypi_0
textblob=0.17.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.12=h27826a3_0
tokenizers=0.12.1=pypi_0
torch=1.12.0=pypi_0
torchmetrics=0.9.2=pypi_0
torchvision=0.13.0=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.0=pypi_0
traitlets=5.3.0=pypi_0
transformers=4.22.2=pypi_0
typing-extensions=4.3.0=pypi_0
urllib3=1.26.9=pypi_0
utils=1.0.1=pypi_0
waitress=2.1.2=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.6.1=pypi_0
xlrd=1.2.0=pypi_0
xlsxwriter=3.0.3=pyhd8ed1ab_0
xxhash=3.0.0=pypi_0
xz=5.2.5=h516909a_1
yarl=1.7.2=pypi_0
zipp=3.8.0=pypi_0
zlib=1.2.12=h166bdaf_1
